Optimize get_etexts for simple metadata extractors
Previously, a call to `get_etexts(A, B)` retrieved the set of triples
`S1 = {(s, p, o) | p == A}` and then performed a linear scan over this set to
remove all the triples for which `o != B`.

With this commit, the same call directly retrieves the much smaller set
`S2 = {(s, p, o) | p == A && o == B}` which makes the function essentially
instantaneous.

Fixes #17