Introduce a Metadata Cache Manager
Encapsulate the metadata cache into its own class, allowing package
users to have more fine-grained control over the cache. This includes
things like it's location and type.

This is almost entirely backwards compatible with the notable exception
that the cache must be manually populated now. A multi-hour process to
populate the cache should be something a user explicitly requests.