Add rdflib-sqlalchemy to Travis config
To allow Travis to test the SQLAlchemy+sqlite backend engine we need
it to install rdflib-sqlalchemy. This RDFLib plug-in isn't available
in pypi so install it directly from github.

Note that we don't make this a package requirement as folks may not
want to pull in SQLAlchemy if they aren't planning on using it.