Add upper version limit for rdflib (#131)
rdflib 5.0.0 reworked some amount of their interface, which breaks things here. Minimally, SPARQLWrapper is no longer used,
and a requests based wrapper internal to rdflib is now used, which breaks compatibility with older RDF stores that do not follow
the SPARQL Protocol 1.1 spec. Setting this to minimize the break here, with a look to move to rdflib 5 in the next major version
bump of gutenberg itself.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>